% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assignments.R
\name{plot_assignments}
\alias{plot_assignments}
\title{Plot histogram names length}
\usage{
plot_assignments(codeExplorer, ...)
}
\arguments{
\item{codeExplorer}{an object of class codeExplorer to be plotted}

\item{...}{other codeExplorers to be plotted}
}
\description{
histgram of unique names
}
