% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_code_PCA.R
\name{plot_codePCA}
\alias{plot_codePCA}
\title{PCA of codes}
\usage{
plot_codePCA(codeExplorer, ..., variables = c("libraries", "functions",
  "mean_length_var", "naming"))
}
\arguments{
\item{codeExplorer}{an object of class codeExplorer to be plotted}

\item{...}{other codeExplorers to be plotted}

\item{variables}{vector of variables to include}
}
\description{
Principal Component Analysis
}
