% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_functions}
\alias{plot_functions}
\title{Plot of used functions}
\usage{
plot_functions(codeExplorer, ..., position = "stack")
}
\arguments{
\item{codeExplorer}{an object of class codeExplorer to be plotted}

\item{...}{other codeExplorers to be plotted}

\item{position}{position of bars}
}
\description{
plot
}
