% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_libraries.R
\name{plot_libraries}
\alias{plot_libraries}
\title{Plot of used libraries}
\usage{
plot_libraries(codeExplorer, ..., position = "stack", plotly = FALSE)
}
\arguments{
\item{codeExplorer}{an object of class codeExplorer to be plotted}

\item{...}{other codeExplorers to be plotted}

\item{position}{position of bars}

\item{ploly}{indicates whenever plotly version should be plotted}
}
\description{
plot
}
