% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_names_hist.R
\name{plot_names_hist}
\alias{plot_names_hist}
\title{Plot histogram names length}
\usage{
plot_names_hist(codeExplorer, ...)
}
\arguments{
\item{codeExplorer}{an object of class codeExplorer to be plotted}

\item{...}{other codeExplorers to be plotted}
}
\description{
histgram of unique names
}
