% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_naming_convention.R
\name{plot_naming_convention}
\alias{plot_naming_convention}
\title{Plot of naming convention functions}
\usage{
plot_naming_convention(codeExplorer, ...)
}
\arguments{
\item{codeExplorer}{an object of class codeExplorer to be plotted}

\item{...}{other codeExplorers to be plotted}
}
\description{
plot
}
