% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_codeExplorer.R
\name{print.codeExplorer}
\alias{print.codeExplorer}
\title{Prints code Summary}
\usage{
\method{print}{codeExplorer}(x, ...)
}
\arguments{
\item{x}{a codeExplainer object created with the `summarize_code` function}

\item{...}{other parameters}
}
\description{
Prints code Summary
}
