% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_code.R
\name{summarize_code}
\alias{summarize_code}
\title{Summary of the code}
\usage{
summarize_code(code, ..., label = NULL, code_names = NULL)
}
\arguments{
\item{code}{Vector of stings, each code line should be a separate string.}

\item{...}{additional vectors with code codes}

\item{label}{character - the name of the explorer. By default it's extracted from the first line of code.}

\item{code_names}{vector of names of codes}
}
\value{
object of class codeExplorer
}
\description{
summarize_code extracts libraries, functions and variables used in code.
}
